StreamCars: Grafische Oberfläche so umgestellt, dass es Graphen mit Zyklen anzeigen kann
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@6719 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: fe7d5a3bdb6bfb0b34fca458b4e1a315b528f887

+++ rcp/de.uniol.inf.is.odysseus.rcp.viewer.view/src/de/uniol/inf/is/odysseus/rcp/viewer/view/render/impl/RenderRange.java
@@ -15,7 +15,7 @@ public final class RenderRange {
 	}
 	
 	public boolean contains( Vector v ) {
-		return contains( v.getX(), v.getY() );
+		return contains( (int)v.getX(), (int)v.getY() );
 	}
 	
 	public boolean intersects (int x, int y, int width, int height) {
@@ -24,6 +24,6 @@ public final class RenderRange {
 	}
 	
 	public boolean intersects( Vector start, int width, int height ) {
-		return intersects( start.getX(), start.getY(), width, height );
+		return intersects( (int)start.getX(), (int)start.getY(), width, height );
 	}
 }
